% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{add_weighted_efficacy}
\alias{add_weighted_efficacy}
\title{Extract weighted efficacies}
\usage{
add_weighted_efficacy(
  dose_number,
  infection_efficacy,
  disease_efficacy,
  delay_dose1,
  delay_dose2
)
}
\arguments{
\item{dose_number}{Dose number: 1 or 2}

\item{infection_efficacy}{Vector of length 2 of infection efficacy for first and second dose}

\item{disease_efficacy}{Vector of length 2 of disease efficacy for first and second dose}

\item{delay_dose1}{Delay between receipt of first dose and protection from first dose (duration in v1v)..}

\item{delay_dose2}{Delay between receipt of second dose and protection from second dose.}
}
\description{
Extract weighted efficacies
}
