% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{administer_first_dose}
\alias{administer_first_dose}
\title{Administer first doses}
\usage{
administer_first_dose(ad, dose_times, t)
}
\arguments{
\item{ad}{Number of doses to administer in each age group}

\item{dose_times}{Dose monitoring data}

\item{t}{Current time}
}
\description{
Administer first doses
}
