% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{administer_second_dose}
\alias{administer_second_dose}
\title{Administer second doses}
\usage{
administer_second_dose(ad, dose_times, dose_period, t)
}
\arguments{
\item{ad}{Number of doses to administer in each age group}

\item{dose_times}{Dose monitoring data}

\item{dose_period}{Days between dose 1 and dose 2}

\item{t}{Current time}
}
\description{
Administer second doses
}
