% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{assign_doses}
\alias{assign_doses}
\title{Assign N doses to age groups based on weightings by number of people eligible to be vaccinated}
\usage{
assign_doses(doses, target_pop)
}
\arguments{
\item{doses}{Total available doses}

\item{target_pop}{Number of people eligible to be vaccinated}
}
\description{
Assign N doses to age groups based on weightings by number of people eligible to be vaccinated
}
