% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{beta_est_infectiousness}
\alias{beta_est_infectiousness}
\title{Estimate beta parameter for explicit model}
\usage{
beta_est_infectiousness(
  dur_IMild,
  dur_ICase,
  prob_hosp,
  rel_infectiousness,
  mixing_matrix,
  R0
)
}
\arguments{
\item{dur_IMild}{Duration of mild infectiousness (days)}

\item{dur_ICase}{Delay between symptom onset and requiring hospitalisation (days)}

\item{prob_hosp}{Probability of hospitilisation by ages}

\item{rel_infectiousness}{Relative infectiousness of age categories relative
to maximum infectiousness age category}

\item{mixing_matrix}{Mixing matrix}

\item{R0}{Basic reproduction number}
}
\value{
Beta parameter
}
\description{
Estimate beta parameter for explicit model
}
