% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{collapse_age}
\alias{collapse_age}
\title{Collapse age groups}
\usage{
collapse_age(x)
}
\arguments{
\item{x}{age-disaggregated odin output matrix}
}
\value{
age-aggregated output matrix
}
\description{
Collapse age groups
}
