% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{coverage}
\alias{coverage}
\title{Estimate coverage by age group}
\usage{
coverage(dose_times, dose_number)
}
\arguments{
\item{dose_times}{Dose monitoring data}

\item{dose_number}{Dose number: 1 or 2}
}
\description{
Estimate coverage by age group
}
