% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_durations}
\alias{default_durations}
\title{Return the default hospital durations for modelling defined in \code{squire}
For more info see \href{squire parameters vignette}{https://mrc-ide.github.io/squire/articles/parameters.html}}
\usage{
default_durations()
}
\value{
list of default durations
}
\description{
Return the default hospital durations for modelling defined in \code{squire}
For more info see \href{squire parameters vignette}{https://mrc-ide.github.io/squire/articles/parameters.html}
}
