% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_probs}
\alias{default_probs}
\title{Return the default probabilities for modelling defined in \code{squire}
For more info see \href{squire parameters vignette}{https://mrc-ide.github.io/squire/articles/parameters.html}}
\usage{
default_probs()
}
\value{
list of default probabilities
}
\description{
Return the default probabilities for modelling defined in \code{squire}
For more info see \href{squire parameters vignette}{https://mrc-ide.github.io/squire/articles/parameters.html}
}
