% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{eligable_for_second}
\alias{eligable_for_second}
\title{Identify those eligible for a second dose}
\usage{
eligable_for_second(dose_times, t, dose_period)
}
\arguments{
\item{dose_times}{Dose monitoring data}

\item{t}{Current time}

\item{dose_period}{Days between dose 1 and dose 2}
}
\description{
Those eligible are those who have had a first dose more than a given period prior and
who have not yet received a second dose.
}
