% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{extract_dose_number}
\alias{extract_dose_number}
\title{Extract dose numbers administers}
\usage{
extract_dose_number(dose_times, maxt)
}
\arguments{
\item{dose_times}{Dose monitoring data}

\item{maxt}{Max time}
}
\description{
Extract dose numbers administers
}
