% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format}
\alias{format}
\title{Format vaccine model output}
\usage{
format(
  x,
  compartments = c("S", "E", "IMild", "ICase", "IICU", "IHospital", "IRec", "R", "D"),
  summaries = c("N", "hospitalisations", "hospital_demand", "hospital_occupancy",
    "ICU_demand", "ICU_occupancy", "vaccines", "unvaccinated", "vaccinated",
    "priorvaccinated", "infections", "deaths"),
  reduce_age = TRUE,
  date_0 = NULL,
  replicate = 1
)
}
\arguments{
\item{x}{squire_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}, or sub-compartment names, e.g. \code{c("S", "E1", "E2")}}

\item{summaries}{Vector of summary names, which may be:
\itemize{
      \item{"deaths"}{ Deaths per day }
      \item{"infections"}{ Infections per day. New infections (note this is currently a slightly different definitionto the main Squire mode)}
      \item{"hospitilisations"}{ Hospitalisations per day (Note this takes into account hospital capacity)}
      \item{"hospital_occupancy"}{ Occupied Hospital Beds }
      \item{"ICU_occupancy"}{ Occupied ICU Beds }
      \item{"hospital_demand}{ Required Hospital Beds }
      \item{"ICU_demand}{ Required ICU Beds }
      \item{"vaccinated"}{ Vaccines administered per day}
      }}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{date_0}{Date of time 0 (e.g. "2020-03-01"), if specified a date column will be added}

\item{replicate}{Which replicate is being formatted. Default = 1}
}
\value{
Formatted long data.frame
}
\description{
Take raw odin vaccine model output and formats in long format with the option to select
variables and summarise over age groups. Output variables are ordered as in argument ordering.
}
