% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritisation_matrix.R
\name{p_all}
\alias{p_all}
\title{Prioritise all}
\usage{
p_all(max_coverage)
}
\arguments{
\item{max_coverage}{Maximum coverage in any one age group}
}
\value{
Prioritisation coverage matrix
}
\description{
Prioritise all
}
