% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritisation_matrix.R
\name{p_elderly}
\alias{p_elderly}
\title{Prioritise Elderly}
\usage{
p_elderly(max_coverage)
}
\arguments{
\item{max_coverage}{Maximum coverage in any one age group}
}
\value{
Prioritisation coverage matrix
}
\description{
Prioritise Elderly
}
