% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritisation_matrix.R
\name{p_risk_elderly_working_children}
\alias{p_risk_elderly_working_children}
\title{Prioritise Working age risk group > Elderly > working age > Children}
\usage{
p_risk_elderly_working_children(max_coverage, risk_proportion)
}
\arguments{
\item{max_coverage}{Maximum coverage in any one age group}

\item{risk_proportion}{Proportion of working age individuals to prioritise if "Risk Elderly Working Children" is choosen}
}
\value{
Prioritisation coverage matrix
}
\description{
Prioritise Working age risk group > Elderly > working age > Children
}
