% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritisation_matrix.R
\name{p_working_elderly_children}
\alias{p_working_elderly_children}
\title{Prioritise Working > Elderly > Children}
\usage{
p_working_elderly_children(max_coverage)
}
\arguments{
\item{max_coverage}{Maximum coverage in any one age group}
}
\value{
Prioritisation coverage matrix
}
\description{
Prioritise Working > Elderly > Children
}
