% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_odin}
\alias{run_odin}
\title{Run the odin model directly}
\usage{
run_odin(pars, mod_gen = vaccine)
}
\arguments{
\item{pars}{a vaccine_parameters object}

\item{mod_gen}{an odin model generation function (either odin or odin.js)}
}
\value{
Simulation output
}
\description{
Run the odin model directly
}
