% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritisation_matrix.R
\name{strategy_matrix}
\alias{strategy_matrix}
\title{Prioritisation matrix}
\usage{
strategy_matrix(strategy, max_coverage = 0.8, risk_proportion = 0)
}
\arguments{
\item{strategy}{Prioritisation strategy
\itemize{
 \item{"All"}{"Target all age groups"}
 \item{"Elderly"}{"Prioritise by most elderly first, working to younger age groups at each step"}
 \item{"Working Elderly Children"}{"Prioritise working ages followed by elderly followed by children"}
 \item{"Risk Elderly Working Children"}{"Prioritise a sub-population of thoe working age, followed by elderly, followed by remiaining working age followed by children"}
}}

\item{max_coverage}{Maximum coverage in any one age group}

\item{risk_proportion}{Proportion of working age individuals to prioritise if "Risk Elderly Working Children" is choosen}
}
\value{
Prioritisation coverage matrix
}
\description{
Choose from one of four preset prioritisation matrices.
}
