% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{target_pop}
\alias{target_pop}
\title{Target number of individual in each age group to vaccinate}
\usage{
target_pop(
  dose_number,
  dose_times,
  prioritisation,
  t,
  dose_period,
  d2_prioritise
)
}
\arguments{
\item{dose_number}{Dose number: 1 or 2}

\item{dose_times}{Dose monitoring data}

\item{prioritisation}{Current row of the prioritisation matrix}

\item{t}{Current time}

\item{dose_period}{Days between dose 1 and dose 2}

\item{d2_prioritise}{Boolean vector indicating which age groups to prioritise for dose 2}
}
\description{
Target number of individual in each age group to vaccinate
}
