% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{weighted_efficacy}
\alias{weighted_efficacy}
\title{Estimate weighted efficacies}
\usage{
weighted_efficacy(
  iso3c,
  N,
  maxt,
  doses_per_day,
  dose_period,
  v1v2,
  prioritisation_matrix,
  d2_prioritise,
  infection_efficacy,
  disease_efficacy
)
}
\arguments{
\item{iso3c}{Country iso3c code}

\item{N}{Number to simulate. Country population is rescaled to this number}

\item{maxt}{Max time}

\item{doses_per_day}{Vector of available doses per day}

\item{dose_period}{Days between dose 1 and dose 2}

\item{v1v2}{Duration in v1v2 (vaccinated but not protected)}

\item{prioritisation_matrix}{Prioritisation matrix}

\item{d2_prioritise}{Boolean vector indicating which age groups are prioritised to receive 2nd dose}

\item{infection_efficacy}{Vector of length 2 of infection efficacy for first and second dose}

\item{disease_efficacy}{Vector of length 2 of disease efficacy for first and second dose}
}
\value{
Age disaggregated dose numbers and efficacies over time.
}
\description{
Estimate weighted age-disaggregated efficacies and numbers of doses over time
to facilitate an approximation of 2 doses with varying 1st and 2nd dose efficacy.
Current assumptions are that the duration in v1 and v2 (vaccinated but not protected)
is fixed and that the number individuals that cannot be vaccinated (due to being
in the hospitalised flow) does not substantivly impact vaccine roll out rates. It is
recommended to run with reasonably large \code{N} to ensure that there are enough
individuals assigned to each age group to get stable results.
}
