% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_efficacy.R
\name{add_weighted_efficacy}
\alias{add_weighted_efficacy}
\title{Extract weighted efficacies}
\usage{
add_weighted_efficacy(dose_number, infection_efficacy, disease_efficacy, v1v2)
}
\arguments{
\item{dose_number}{Dose number: 1 or 2}

\item{infection_efficacy}{Vector of length 2 of infection efficacy for first and second dose}

\item{disease_efficacy}{Vector of length 2 of disease efficacy for first and second dose}

\item{v1v2}{Duration in v1v2 (vaccinated but not protected)}
}
\description{
Extract weighted efficacies
}
