% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Check and set up initial values for vaccine model}
\usage{
init(population, seeding_cases)
}
\arguments{
\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{seeding_cases}{Initial number of cases seeding the epidemic}
}
\value{
Checked initial values data.frame
}
\description{
Check and set up initial values for vaccine model
}
