% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Check and set up initial values for vaccine model}
\usage{
init(population, seeding_cases, seeding_age_order = NULL)
}
\arguments{
\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{seeding_cases}{Initial number of cases seeding the epidemic}

\item{seeding_age_order}{Vector specifying the order in which seeds are allocated to ages.
If NULL, seeds are distributed randomly within working ages. If specified, must be a vector
of length 17 specifying the order seeds are allocated, e.g. 1:17 will allocate first seed
to the youngest age group, then the second youngest and so on. Default = NULL}
}
\value{
Checked initial values data.frame
}
\description{
Check and set up initial values for vaccine model
}
